innovlead-education-bot/requirements.txt
# InnovLead Education Bot - Python Dependencies

# ============================================================================
# CORE BOT FRAMEWORK
# ============================================================================
python-telegram-bot==20.7
python-dotenv==1.0.0

# ============================================================================
# AI PROVIDERS
# ============================================================================
# Anthropic Claude (recommended)
anthropic==0.18.1

# OpenAI (alternative)
openai==1.12.0

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy==2.0.25
alembic==1.13.1

# ============================================================================
# DATA VALIDATION & SETTINGS
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================================================
# HTTP CLIENTS & API INTEGRATION
# ============================================================================
requests==2.31.0
httpx==0.26.0

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
loguru==0.7.2

# ============================================================================
# DATE & TIME UTILITIES
# ============================================================================
python-dateutil==2.8.2
pytz==2024.1

# ============================================================================
# REPORTING & DATA EXPORT
# ============================================================================
# PDF generation for certificates
reportlab==4.0.9

# Excel reports for CRA compliance
openpyxl==3.1.2

# JSON handling
ujson==5.9.0

# ============================================================================
# MARKDOWN & TEXT PROCESSING
# ============================================================================
markdown==3.5.2
beautifulsoup4==4.12.3

# ============================================================================
# ASYNC UTILITIES
# ============================================================================
aiofiles==23.2.1
asyncio==3.4.3

# ============================================================================
# SECURITY & ENCRYPTION
# ============================================================================
cryptography==42.0.0

# ============================================================================
# RATE LIMITING
# ============================================================================
ratelimit==2.2.1

# ============================================================================
# TESTING (DEVELOPMENT)
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# ============================================================================
# CODE QUALITY (DEVELOPMENT)
# ============================================================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# ============================================================================
# OPTIONAL: N8N INTEGRATION
# ============================================================================
# Uncomment if using n8n workflows
# n8n-client==0.1.0
